Revert "resolving a (weird?) merge conflict on the branch?"
This reverts commit 815ef304b246cab3727b326608b5f86dd6922461, reversing
changes made to 4181948e290014b390ee44ccb2e7d4fefdfb3f63.
