+++ src/main/java/edu/harvard/iq/dataverse/api/BatchServiceBean.java
@@ -5,6 +5,7 @@ import edu.harvard.iq.dataverse.Dataverse;
 import edu.harvard.iq.dataverse.DataverseServiceBean;
 import edu.harvard.iq.dataverse.api.imports.ImportException;
 import edu.harvard.iq.dataverse.api.imports.ImportUtil;
+import edu.harvard.iq.dataverse.authorization.users.AuthenticatedUser;
 import edu.harvard.iq.dataverse.engine.command.DataverseRequest;
 import java.io.File;
 import java.io.FileWriter;
@@ -106,6 +107,5 @@ public class BatchServiceBean {
         }
         return status;
     }
-    
 
 }
