+++ src/main/java/edu/harvard/iq/dataverse/api/Datasets.java
@@ -57,9 +57,6 @@ import javax.ws.rs.Produces;
 import javax.ws.rs.QueryParam;
 import javax.ws.rs.core.MediaType;
 import javax.ws.rs.core.Response;
-import static edu.harvard.iq.dataverse.api.AbstractApiBean.errorResponse;
-import static edu.harvard.iq.dataverse.api.AbstractApiBean.errorResponse;
-import static edu.harvard.iq.dataverse.api.AbstractApiBean.errorResponse;
 
 @Path("datasets")
 public class Datasets extends AbstractApiBean {
@@ -315,7 +312,7 @@ public class Datasets extends AbstractApiBean {
             boolean updateDraft = ds.getLatestVersion().isDraft();
             DatasetVersion managedVersion = execCommand( updateDraft
                                                              ? new UpdateDatasetVersionCommand(req, incomingVersion)
-                                                             : new  CreateDatasetVersionCommand(req, ds, incomingVersion));
+                                                             : new CreateDatasetVersionCommand(req, ds, incomingVersion));
             return okResponse( json(managedVersion) );
                     
         } catch (JsonParseException ex) {
