+++ src/main/java/edu/harvard/iq/dataverse/api/Search.java
@@ -27,11 +27,9 @@ import javax.ejb.EJB;
 import javax.json.Json;
 import javax.json.JsonArrayBuilder;
 import javax.json.JsonObjectBuilder;
-import javax.servlet.http.HttpServletResponse;
 import javax.ws.rs.GET;
 import javax.ws.rs.Path;
 import javax.ws.rs.QueryParam;
-import javax.ws.rs.core.Context;
 import javax.ws.rs.core.Response;
 import org.apache.commons.lang.StringUtils;
 
@@ -67,8 +65,7 @@ public class Search extends AbstractApiBean {
             @QueryParam("fq") final List<String> filterQueries,
             @QueryParam("show_entity_ids") boolean showEntityIds,
             @QueryParam("show_api_urls") boolean showApiUrls,
-            @QueryParam("show_my_data") boolean showMyData,
-            @Context HttpServletResponse response
+            @QueryParam("show_my_data") boolean showMyData
     ) {
 
         User user;
@@ -185,7 +182,6 @@ public class Search extends AbstractApiBean {
                  */
                 return errorResponse(Response.Status.BAD_REQUEST, solrQueryResponse.getError());
             }
-            response.setHeader("Access-Control-Allow-Origin", "*");
             return okResponse(value);
         } else {
             return errorResponse(Response.Status.BAD_REQUEST, "q parameter is missing");
