+++ src/main/java/edu/harvard/iq/dataverse/api/imports/ImportServiceBean.java
@@ -185,111 +185,6 @@ public class ImportServiceBean {
         }
     }
 
-    public Dataset doImportHarvestedDataset(DataverseRequest dataverseRequest, Dataverse owner, String metadataFormat, File metadataFile, PrintWriter cleanupLog) throws ImportException, IOException {
-        Dataset importedDataset = null;
-
-        DatasetDTO dsDTO = null;
-
-        if ("ddi".equals(metadataFormat)) {
-            try {
-                String xmlToParse = new String(Files.readAllBytes(metadataFile.toPath()));
-                dsDTO = importDDIService.doImport(ImportType.HARVEST, xmlToParse);
-            } catch (XMLStreamException e) {
-                throw new ImportException("XMLStreamException" + e);
-            }
-        } // TODO: handle all supported formats; via plugins, probably
-        // (and if the format is already JSON - handle that too!
-        else {
-            throw new ImportException("Unsupported import metadata format: " + metadataFormat);
-        }
-
-        // convert DTO to Json, 
-        Gson gson = new GsonBuilder().setPrettyPrinting().create();
-        String json = gson.toJson(dsDTO);
-        JsonReader jsonReader = Json.createReader(new StringReader(json));
-        JsonObject obj = jsonReader.readObject();
-        //and call parse Json to read it into a dataset   
-        try {
-            JsonParser parser = new JsonParser(datasetfieldService, metadataBlockService, settingsService);
-            parser.setLenient(true);
-            Dataset ds = parser.parseDataset(obj);
-
-            // For ImportType.NEW, if the metadata contains a global identifier, and it's not a protocol
-            // we support, it should be rejected.
-            // (TODO: ! - add some way of keeping track of supported protocols!)
-            //if (ds.getGlobalId() != null && !ds.getProtocol().equals(settingsService.getValueForKey(SettingsServiceBean.Key.Protocol, ""))) {
-            //    throw new ImportException("Could not register id " + ds.getGlobalId() + ", protocol not supported");
-            //}
-            ds.setOwner(owner);
-            ds.getLatestVersion().setDatasetFields(ds.getLatestVersion().initDatasetFields());
-
-            // Check data against required contraints
-            List<ConstraintViolation> violations = ds.getVersions().get(0).validateRequired();
-            if (!violations.isEmpty()) {
-                // For migration and harvest, add NA for missing required values
-                for (ConstraintViolation v : violations) {
-                    DatasetField f = ((DatasetField) v.getRootBean());
-                    f.setSingleValue(DatasetField.NA_VALUE);
-                }
-            }
-
-            // Check data against validation constraints
-            // If we are migrating and "scrub migration data" is true we attempt to fix invalid data
-            // if the fix fails stop processing of this file by throwing exception
-            Set<ConstraintViolation> invalidViolations = ds.getVersions().get(0).validate();
-            ValidatorFactory factory = Validation.buildDefaultValidatorFactory();
-            Validator validator = factory.getValidator();
-            if (!invalidViolations.isEmpty()) {
-                for (ConstraintViolation v : invalidViolations) {
-                    DatasetFieldValue f = ((DatasetFieldValue) v.getRootBean());
-                    boolean fixed = false;
-                    boolean converted = false;
-                    // TODO: Is this scrubbing something we want to continue doing? 
-                    if (settingsService.isTrueForKey(SettingsServiceBean.Key.ScrubMigrationData, false)) {
-                        fixed = processMigrationValidationError(f, cleanupLog, metadataFile.getName());
-                        converted = true;
-                        if (fixed) {
-                            Set<ConstraintViolation<DatasetFieldValue>> scrubbedViolations = validator.validate(f);
-                            if (!scrubbedViolations.isEmpty()) {
-                                fixed = false;
-                            }
-                        }
-                    }
-                    if (!fixed) {
-                        String msg = "Data modified - File: " + metadataFile.getName() + "; Field: " + f.getDatasetField().getDatasetFieldType().getDisplayName() + "; "
-                                + "Invalid value:  '" + f.getValue() + "'" + " Converted Value:'" + DatasetField.NA_VALUE + "'";
-                        cleanupLog.println(msg);
-                        f.setValue(DatasetField.NA_VALUE);
-
-                    }
-                }
-            }
-
-            Dataset existingDs = datasetService.findByGlobalId(ds.getGlobalId());
-
-            if (existingDs != null) {
-                // For harvested datasets, there should always only be one version.
-                // We will replace the current version with the imported version.
-                if (existingDs.getVersions().size() != 1) {
-                    throw new ImportException("Error importing Harvested Dataset, existing dataset has " + existingDs.getVersions().size() + " versions");
-                }
-                engineSvc.submit(new DestroyDatasetCommand(existingDs, dataverseRequest));
-                importedDataset = engineSvc.submit(new CreateDatasetCommand(ds, dataverseRequest, false, ImportType.HARVEST));
-
-            } else {
-                importedDataset = engineSvc.submit(new CreateDatasetCommand(ds, dataverseRequest, false, ImportType.HARVEST));
-            }
-
-        } catch (JsonParseException ex) {
-            logger.log(Level.INFO, "Error parsing datasetVersion: {0}", ex.getMessage());
-            throw new ImportException("Error parsing datasetVersion: " + ex.getMessage(), ex);
-        } catch (CommandException ex) {
-            logger.log(Level.INFO, "Error excuting Create dataset command: {0}", ex.getMessage());
-            throw new ImportException("Error excuting dataverse command: " + ex.getMessage(), ex);
-        }
-        return importedDataset;
-    }
-    
     public JsonObjectBuilder doImport(DataverseRequest dataverseRequest, Dataverse owner, String xmlToParse, String fileName, ImportType importType, PrintWriter cleanupLog) throws ImportException, IOException {
 
         String status = "";
